+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+- Added support wildcards in ```excludedFilenames``` and ```excludedDirectories``` properties
+
 ## [1.1.0] - 2026-03-26
 
 ### Added